Make options tests work
