+++ tests/unit/test_options.py
@@ -1,39 +1,19 @@
 import os
 
-import pip.configuration
 import pytest
-from pip import cmdoptions, main
-from pip.basecommand import Command
-from pip.commands import commands_dict as commands
+
+import pip.configuration
+from pip import main
+
+from tests.lib.options_helpers import AddFakeCommandMixin
 
 
-class FakeCommand(Command):
-    name = 'fake'
-    summary = name
-
-    def main(self, args):
-        index_opts = cmdoptions.make_option_group(
-            cmdoptions.index_group,
-            self.parser,
-        )
-        self.parser.add_option_group(index_opts)
-        return self.parse_args(args)
-
-
-class TestOptionPrecedence(object):
+class TestOptionPrecedence(AddFakeCommandMixin):
     """
     Tests for confirming our option precedence:
-         cli -> environment -> subcommand config -> global config -> option
-         defaults
+        cli -> environment -> subcommand config -> global config -> option
+        defaults
     """
-
-    def setup(self):
-        self.environ_before = os.environ.copy()
-        commands[FakeCommand.name] = FakeCommand
-
-    def teardown(self):
-        os.environ = self.environ_before
-        commands.pop(FakeCommand.name)
 
     def get_config_section(self, section):
         config = {
@@ -102,53 +82,8 @@ class TestOptionPrecedence(object):
         options, args = main(['fake', '--timeout', '-2'])
         assert options.timeout == -2
 
-    def test_environment_override_config(self, monkeypatch):
-        """
-        Test an environment variable overrides the config file
-        """
-        monkeypatch.setattr(
-            pip.configuration.Configuration,
-            "_get_config_section",
-            self.get_config_section,
-        )
-        os.environ['PIP_TIMEOUT'] = '-1'
-        options, args = main(['fake'])
-        assert options.timeout == -1
 
-    def test_commmand_config_override_global_config(self, monkeypatch):
-        """
-        Test that command config overrides global config
-        """
-        monkeypatch.setattr(
-            pip.configuration.Configuration,
-            "_get_config_section",
-            self.get_config_section,
-        )
-        options, args = main(['fake'])
-        assert options.timeout == -2
-
-    def test_global_config_is_used(self, monkeypatch):
-        """
-        Test that global config is used
-        """
-        monkeypatch.setattr(
-            pip.configuration.Configuration,
-            "_get_config_section",
-            self.get_config_section_global,
-        )
-        options, args = main(['fake'])
-        assert options.timeout == -3
-
-
-class TestOptionsInterspersed(object):
-
-    def setup(self):
-        self.environ_before = os.environ.copy()
-        commands[FakeCommand.name] = FakeCommand
-
-    def teardown(self):
-        os.environ = self.environ_before
-        commands.pop(FakeCommand.name)
+class TestOptionsInterspersed(AddFakeCommandMixin):
 
     def test_general_option_after_subcommand(self):
         options, args = main(['fake', '--timeout', '-1'])
@@ -167,18 +102,10 @@ class TestOptionsInterspersed(object):
             main(['--find-links', 'F1', 'fake'])
 
 
-class TestGeneralOptions(object):
+class TestGeneralOptions(AddFakeCommandMixin):
 
     # the reason to specifically test general options is due to the
     # extra processing they receive, and the number of bugs we've had
-
-    def setup(self):
-        self.environ_before = os.environ.copy()
-        commands[FakeCommand.name] = FakeCommand
-
-    def teardown(self):
-        os.environ = self.environ_before
-        commands.pop(FakeCommand.name)
 
     def test_require_virtualenv(self):
         options1, args1 = main(['--require-virtualenv', 'fake'])
@@ -274,4 +201,9 @@ class TestOptionsConfigFiles(object):
         )
         monkeypatch.setattr(os.path, 'exists', lambda filename: True)
         cp = pip.configuration.Configuration(isolated=False)
-        assert len(cp._get_config_files()) == 4
+
+        files = []
+        for _, val in cp._get_config_files():
+            files.extend(val)
+
+        assert len(files) == 4
